improvements for asynchronous clustering

+++ src/routes/api/index/nodes/[node_id]/+server.ts
@@ -1,14 +1,19 @@
-import { PUBLIC_INDEX_URL, PUBLIC_TOOLS_URL } from '$env/static/public';
+import { PUBLIC_TOOLS_URL } from '$env/static/public';
 import { json, type RequestHandler } from '@sveltejs/kit';
 
-export const GET: RequestHandler = async ({ params }) => {
+export const GET: RequestHandler = async ({ params, url }) => {
 	try {
 		const { node_id } = params;
 		if (!node_id) {
 			return json({ error: 'Missing node_id', success: false }, { status: 400 });
 		}
 
-		const response = await fetch(`${PUBLIC_INDEX_URL}/v2/nodes/${node_id}`);
+		const source_index_url = url.searchParams.get('source_index_url');
+		if (!source_index_url) {
+			return json({ error: 'Missing source_index_url', success: false }, { status: 400 });
+		}
+
+		const response = await fetch(`${source_index_url}/v2/nodes/${node_id}`);
 
 		const data = await response.json();
 		if (!response.ok) {
@@ -30,8 +35,7 @@ export const GET: RequestHandler = async ({ params }) => {
 		);
 	}
 };
-
-export const POST: RequestHandler = async ({ params }) => {
+export const POST: RequestHandler = async ({ params, request }) => {
 	try {
 		const { node_id } = params;
 
@@ -39,9 +43,15 @@ export const POST: RequestHandler = async ({ params }) => {
 			return json({ error: 'Missing node_id', success: false }, { status: 400 });
 		}
 
+		const { source_index_url: sourceIndexUrl } = await request.json();
+
+		if (!sourceIndexUrl) {
+			return json({ error: 'Missing source_index_url', success: false }, { status: 400 });
+		}
+
 		const profileUrl = `${PUBLIC_TOOLS_URL}/profiles/${node_id}`;
 
-		const response = await fetch(`${PUBLIC_INDEX_URL}/v2/nodes`, {
+		const response = await fetch(`${sourceIndexUrl}/v2/nodes`, {
 			method: 'POST',
 			headers: {
 				'Content-Type': 'application/json'
@@ -70,7 +80,7 @@ export const POST: RequestHandler = async ({ params }) => {
 	}
 };
 
-export const DELETE: RequestHandler = async ({ params }) => {
+export const DELETE: RequestHandler = async ({ params, request }) => {
 	try {
 		const { node_id } = params;
 
@@ -78,7 +88,12 @@ export const DELETE: RequestHandler = async ({ params }) => {
 			return json({ error: 'Missing node_id', success: false }, { status: 400 });
 		}
 
-		const response = await fetch(`${PUBLIC_INDEX_URL}/v2/nodes/${node_id}`, {
+		const { source_index_url: sourceIndexUrl } = await request.json();
+		if (!sourceIndexUrl) {
+			return json({ error: 'Missing source_index_url', success: false }, { status: 400 });
+		}
+
+		const response = await fetch(`${sourceIndexUrl}/v2/nodes/${node_id}`, {
 			method: 'DELETE'
 		});
